app/requirements.txt
flask==0.10.1
redis
mysql-connector-python==8.0.12